@@ -13,15 +13,15 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/schematics": "^19.1.8",
-    "@angular/animations": "^18.2.13",
-    "@angular/common": "^18.2.13",
-    "@angular/compiler": "^18.2.13",
-    "@angular/core": "^18.2.13",
-    "@angular/forms": "^18.2.13",
-    "@angular/http": "^7.2.16",
-    "@angular/platform-browser": "^18.2.13",
-    "@angular/platform-browser-dynamic": "^18.2.13",
-    "@angular/router": "^18.2.13",
+    "@angular/animations": "^19.1.7",
+    "@angular/common": "^19.1.7",
+    "@angular/compiler": "^19.1.7",
+    "@angular/core": "^19.1.7",
+    "@angular/forms": "^19.1.7",
+    "@angular/platform-browser": "^19.1.7",
+    "@angular/platform-browser-dynamic": "^19.1.7",
+    "@angular/router": "^19.1.7",
+    "@ngu/carousel": "^19.0.0",
     "@schematics/angular": "^19.1.8",
     "angular-bem": "^2.0.3",
     "angular-in-memory-web-api": "^0.3.2",
@@ -29,16 +29,15 @@
     "hammerjs": "^2.0.8",
     "ng-sidebar": "^7.0.1",
     "ng2-scroll-to": "^1.0.7",
-    "ngx-carousel": "^1.3.5",
-    "rxjs": "^6.6.7",
-    "tslib": "^1.10.0",
-    "zone.js": "^0.14.10"
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.14",
-    "@angular/cli": "18.2.14",
-    "@angular/compiler-cli": "^18.2.13",
-    "@angular/language-service": "^18.2.13",
+    "@angular-devkit/build-angular": "^19.1.8",
+    "@angular/cli": "19.1.8",
+    "@angular/compiler-cli": "^19.1.7",
+    "@angular/language-service": "^19.1.7",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
