@@ -1,25 +1,32 @@
 import { Injectable } from '@angular/core';
-import { Headers, Http } from '@angular/http';
+import { HttpHeaders, HttpClient } from '@angular/common/http';
 
-import 'rxjs/add/operator/toPromise';
+// import 'rxjs/add/operator/toPromise';
+// import { Observable } from 'rxjs'; 
+import { Observable } from 'rxjs'
 
 import { Project } from './../projects/project';
 
-@Injectable()
+@Injectable({ 
+    providedIn: 'root' 
+}) 
+
 export class ProjectService {
     
     private headers = new Headers({'Content-Type': 'application/json'});
-    private projectsUrl = 'api/projects';  // URL to web api
+    // private projectsUrl = 'api/projects';  // URL to web api
+    private apiUrl = 'api/projects'; // Remplacez par l'URL de votre API 
 
     constructor(
-        private http: Http,
+        private http: HttpClient,
         // private backend: BackendService
     ) { }
+
 
     getProjects(): Promise<Project[]> {
         return this.http.get(this.projectsUrl)
             .toPromise()
-            .then(response => response.json().data as Project[])
+            .then(response => response?.json().data as Project[])
             .catch(this.handleError);
     }
 
@@ -31,13 +38,15 @@ export class ProjectService {
     //     return this.heroes;
     // }
 
+
     getProject(id: number): Promise<Project> {
         const url = `${this.projectsUrl}/${id}`;
         return this.http.get(url)
             .toPromise()
-            .then(response => response.json().data as Project[])
+            .then(response => response?.json().data as Project[])
             .catch(this.handleError);
     }
+
 
     private handleError(error: any): Promise<any> {
         console.error('An error occurred', error); // for demo purposes only
