@@ -1,13 +1,16 @@
-import { Component, OnInit } from '@angular/core';
-import { DatePipe }          from '@angular/common';
-import { Title }             from '@angular/platform-browser';
-import { Router }            from '@angular/router';
+import { Component, Inject, OnInit } from '@angular/core';
+import { CommonModule, DatePipe }    from '@angular/common';
+import { Title }                     from '@angular/platform-browser';
+import { RouterOutlet }              from '@angular/router';
+import { SidebarComponent }         from './sidebar/sidebar.component';
 
 
 @Component({
     selector: 'my-app',
+    standalone: true,
+    imports: [RouterOutlet, SidebarComponent, CommonModule],
     templateUrl: './app.component.html',
-    styleUrls: ['./app.component.scss']
+    styleUrl: './app.component.scss'
 })
 
 export class AppComponent implements OnInit {
@@ -16,23 +19,29 @@ export class AppComponent implements OnInit {
     fixedTimezone = '2015-06-15T09:03:01+0900';
     
     private mobile = false;
-    private _opened: boolean = false;
-    private _modeNum: number = 0;
-    private _closeOnClickBackdrop: boolean = true;
-    private _showBackdrop: boolean = true;
 
-    private _MODES: Array<string> = ['over', 'push', 'slide'];
-    
-    private _toggleSidebar(){
-        this._opened = !this._opened;
+    isSidebarCollapsed = true;
+    onSidebarToggle() {
+        this.isSidebarCollapsed = !this.isSidebarCollapsed;
     }
+
+    // PREVIOUS SIDEBAR (??)
+    // private _opened: boolean = false;
+    // private _modeNum: number = 0;
+    // private _closeOnClickBackdrop: boolean = true;
+    // private _showBackdrop: boolean = true;
+    // private _MODES: Array<string> = ['over', 'push', 'slide'];
+    // private _toggleSidebar(){
+    //     this._opened = !this._opened;
+    // }
+    // END PREVIOUS SIDEBAR (??)
 
     public constructor(private titleService: Title ) { }
     public setTitle( newTitle: string) {
         this.titleService.setTitle( newTitle );
     }
 
-    getState(outlet) {
+    getState(outlet: any) {
         return outlet.activatedRouteData.state;
     }
 
