@@ -1,35 +1,35 @@
-import { NgModule }             from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { NgModule }                 from '@angular/core';
+import { RouterModule, Routes }                   from '@angular/router';
 
 import { HomeComponent }            from './../home/home.component';
 import { PresentationComponent }    from './../informations/presentation.component';
 import { CareerComponent }          from './../informations/career.component';
 // import { PageNotFoundComponent }    from './../pageNotFound.component';
 
-const appRoutes: Routes = [
+const routes: Routes = [
     {
         path: 'accueil',
         component: HomeComponent,
-        data: {
-            animation: 'home'
-        }
+        // data: {
+        //     animation: 'home'
+        // }
     },
     {
         path: 'presentation',
         component: PresentationComponent,
-        data: {
-            animation: 'presentation'
-        }
+        // data: {
+        //    animation: 'presentation'
+        // }
     },
     {
         path: 'parcours',
         component: CareerComponent,
-        data: {
-            animation: 'career'
-        }
+        // data: {
+        //     animation: 'career'
+        // }
     },
     {
-        path: '',
+        path: '**',
         redirectTo: '/accueil',
         pathMatch: 'full',
         data: {
@@ -40,7 +40,10 @@ const appRoutes: Routes = [
 ];
 
 @NgModule({
-    imports: [ RouterModule.forRoot(appRoutes/*, { enableTracing: true }*/ /*only for debuging purpose */) ],
-    exports: [ RouterModule ]
+    imports: [ RouterModule.forRoot(routes) ],
+    exports: [ RouterModule ],
+    declarations: []
+    //export const routes: Routes = [ RouterModule ] 
 })
+
 export class AppRoutingModule {}
